requirements.txt
fastapi
httpx
jinja2
google-cloud-logging==3.0.0
google-cloud-pubsub==2.12.0
google-cloud-bigquery==2.34.0
google-cloud-firestore==2.11.1
google-auth==2.6.6
PyYAML==6.0
redis==4.3.4
jsonschema== 4.16.0
requests==2.28.1
retry==0.9.2
backoff==2.2.1
cachetools==5.3.0
data-common @ git+ssh://git@github.com/brutmedia/data-common.git@1.4.3
xminds==0.3.0
spectree==1.0.3